Crate relay_event_normalization

Event normalization and processing.

Modules

breakdowns

Computation of breakdowns from event data.

contexts

Computation and normalization of contexts from event data.

eap

Event normalization and processing for attribute (EAP) based payloads.

nel

Contains helper function for NEL reports.

replay

Validation and normalization of Replay events.

request

Normalization of the Request interface.

span

Span normalization logic.

user_agent

Generate context data from user agent and client hints.

utils

Deprecated. Utilities for extracting common event fields.

Structs

AiOperationTypeMap

A mapping of AI operation types from span.op to gen_ai.operation.type.

BorrowedSpanOpDefaults

Borrowed version of SpanOpDefaults.

BreakdownsConfig

Configuration for computing breakdowns from data in the event.

BuiltinMeasurementKey

Defines a builtin measurement.

ClientHints

The client hint variable names mirror the name of the “SEC-CH” headers.

ClockDriftProcessor

Corrects clock drift based on the sender’s and receivers timestamps.

CombinedMeasurementsConfig

Container for global and project level MeasurementsConfig. The purpose is to handle the merging logic.

EventValidationConfig

Configuration for validate_event.

GeoIpLookup

A geo ip lookup helper based on maxmind db files.

MeasurementsConfig

Configuration for measurements normalization.

ModelCostV2

Version 2 of a mapping of AI model types (like GPT-4) to their respective costs.

ModelCosts

A mapping of AI model types (like GPT-4) to their respective costs.

NormalizationConfig

Configuration for normalize_event.

PerformanceScoreConfig

Defines the performance configuration for the project.

PerformanceScoreProfile

Defines a profile for performance score.

PerformanceScoreWeightedComponent

Defines a weighted component for a performance score.

RawUserAgentInfo

A container housing both the user-agent string and the client hint headers.

RemoveOtherProcessor

Removes unknown, internal and deprecated fields from a payload.

SchemaProcessor

Validates constraints such as empty strings or arrays and invalid characters.

SpanDescriptionRule

The rule describes how span descriptions should be changed.

SpanDescriptionRuleScope

Object containing transaction attributes the rules must only be applied to.

SpanOpDefaultRule

A rule to infer Span::op from other span fields.

SpanOpDefaults

Rules used to infer span.op from other span fields.

SpanOperationsConfig

Configuration to define breakdowns based on span operation name.

TimeWindowSpan

A time window declared by its start and end timestamp.

TimestampProcessor

Ensures an event’s timestamps are not stale.

TransactionNameConfig

Configuration for sanitizing unparameterized transaction names.

TransactionNameRule

The rule describes how transaction name should be changed.

TransactionsProcessor

Rejects transactions based on required fields.

TrimmingProcessor

Limits properties to a maximum size and depth.

Enums

BreakdownConfig

Configuration to define breakdown to be generated based on properties and breakdown type.

RedactionRule

Describes what to do with the matched pattern.

Traits

FromUserAgentInfo

Computes a Context from either a user agent string and client hints.

Functions

apply_transaction_rename_rules

Applies the rule if any found to the transaction name.

browser_from_client_hints

The sec-ch-ua field looks something like this: “Not_A Brand”;v=“99”, “Google Chrome”;v=“109”, “Chromium”;v=“109” The order of the items are randomly shuffled.

is_high_cardinality_sdk

Returns true if the event’s transaction name is known to contain unsanitized values.

is_valid_platform

Returns true if the given platform string is a known platform identifier.

normalize_app_start_spans

Replaces snake_case app start spans op with dot.case op.

normalize_breakdowns

Computes breakdowns for an event based on the given configuration.

normalize_event

Normalizes an event.

normalize_measurements

Ensure only valid measurements are ingested.

normalize_performance_score

Computes performance score measurements for an event.

normalize_transaction_name

Apply parametrization to transaction.

normalize_user_agent

Generates context data from client hints or user agent.

normalize_user_agent_info_generic

Low-level version of normalize_user_agent operating on parts.

set_default_transaction_source

Set a default transaction source if it is missing, but only if the transaction name was extracted as a metrics tag. This behavior makes it possible to identify transactions for which the transaction name was not extracted as a tag on the corresponding metrics, because source == null <=> transaction name == null See relay_server::metrics_extraction::transactions::get_transaction_name.

validate_environment

Given a string checks if the environment name is generally valid.

validate_event

Validates an event.

validate_release

Given a string checks if the release is generally valid.

validate_span

Validates a span.

Type Aliases

GeoIpError

An error in the GeoIpLookup.