relay_event_normalization

Module eap

Event normalization and processing for attribute (EAP) based payloads.

A central place for all modifications/normalizations for attributes.

Modules

time

Time normalization for EAP items.

trace_metric

Normalizations specific to trace metrics.

Structs

TrimmingProcessor

Processor for trimming EAP items (logs, V2 spans).

Functions

attribute_size

Calculates the size of a single attribute.

attributes_size

Calculates the canonical size of Attributes.

normalize_ai

Normalizes AI attributes.

normalize_attribute_names

Normalizes deprecated attributes according to sentry-conventions.

normalize_attribute_types

Normalizes/validates all attribute types.

normalize_attribute_values

Normalizes the values of a set of attributes if present in the span.

normalize_client_address

Normalizes the client address into Attributes.

normalize_dsc

Normalizes the DSC into Attributes.

normalize_received

Adds the received time to the attributes.

normalize_sentry_op

Infers the sentry.op attribute and inserts it into Attributes if not already set.

normalize_span_category

Infers the sentry.category attribute and inserts it into attributes if not already set. The category is derived from the span operation or other span attributes.

normalize_user_agent

Normalizes the user agent/client information into Attributes.

normalize_user_geo

Normalizes the user’s geographical information into Attributes.

value_size

Recursively calculates the size of a Value, using the rules described in attributes_size.

write_legacy_attributes

Double writes sentry conventions attributes into legacy attributes.