relay_event_normalization::eap

Function normalize_attribute_values

pub fn normalize_attribute_values(
    attributes: &mut Annotated<Attributes>,
    http_span_allowed_hosts: &[String],
)

Normalizes the values of a set of attributes if present in the span.

Each span type has a set of important attributes containing the main relevant information displayed in the product-end. For instance, for DB spans, these attributes are db.query.text, db.operation.name, db.collection.name. Previously, V1 spans always held these important values in the description field, however, V2 spans now store these values in their respective attributes based on sentry conventions. This function ports over the SpanV1 normalization logic that was previously in scrub_span_description by creating a set of functions to handle each group of attributes separately.