relay_event_normalization::eap

Function normalize_client_address

pub fn normalize_client_address(
    attributes: &mut Annotated<Attributes>,
    client_ip: Option<IpAddr>,
)

Normalizes the client address into [Attributes].

Infers the client ip from the client information which was provided to Relay, if the SDK indicates the client ip should be inferred by setting it to {{auto}}.

This requires cooperation from SDKs as inferring a client ip only works in non-server environments, where the user/client device is also the device sending the item.