relay_event_normalization

Function validate_event

pub fn validate_event(
    event: &mut Annotated<Event>,
    config: &EventValidationConfig,
) -> ProcessingResult

Validates an event.

Validation consists of performing multiple checks on the payload, based on the given configuration.

The returned ProcessingResult indicates whether the passed event is invalid and thus should be dropped.