relay_event_normalization::span

Module description

Span description scrubbing logic.

Functions

concatenate_host_and_port

Concatenate an optional host and an optional port.

scrub_db_query

Scrubs a DB query string based on relevant attributes within DB spans.

scrub_domain_name

Sanitize a qualified domain string.

scrub_host

Scrub a Host object.

scrub_http

Returns a scrubbed HTTP description of the format: “{method} {scheme}://{domain}” given a method, url, and a list of allowed hosts.

scrub_ipv4

Scrub an IPv4 address.

scrub_ipv6

Scrub an IPv6 address.