relay_event_normalization::span

Module tag_extraction

Logic for persisting items into span.sentry_tags and span.measurements fields. These are then used for metrics extraction.

Statics

HTTP_METHOD_EXTRACTOR_REGEX

Regex with a capture group to extract the HTTP method from a string.

Functions

domain_from_scrubbed_http

Extracts the domain from a scrubbed HTTP description.

domain_from_server_address

Extracts the domain from a server address.

extract_measurements

Copies specific numeric values from span data to span measurements.

extract_segment_span_tags

Extract segment span specific tags and measurements from the event and materialize them into segment span.

extract_span_tags

Extracts tags and measurements from event and spans and materializes them into the spans.

extract_tags

Writes fields into [Span::sentry_tags].

span_op_to_category

Returns the category of a span from its operation. The mapping is available in: https://develop.sentry.dev/sdk/performance/span-operations/

sql_action_from_query

Returns the SQL action from a raw query.

sql_tables_from_query

Returns a sorted, comma-separated list of SQL tables, if any.