relay_event_normalization::span::tag_extraction
Enum SpanTagKey

pub enum SpanTagKey {

    Release,
    User,
    UserID,
    UserIP,
    UserUsername,
    UserEmail,
    Environment,
    Transaction,
    TransactionMethod,
    TransactionOp,
    BrowserName,
    SdkName,
    SdkVersion,
    Platform,
    Mobile,
    DeviceClass,
    OsName,
    Action,
    AIPipelineGroup,
    Category,
    Description,
    Domain,
    RawDomain,
    Group,
    HttpDecodedResponseContentLength,
    HttpResponseContentLength,
    HttpResponseTransferSize,
    ResourceRenderBlockingStatus,
    SpanOp,
    SpanStatus,
    StatusCode,
    System,
    TimeToInitialDisplay,
    TimeToFullDisplay,
    FileExtension,
    MainThread,
    AppStartType,
    ReplayId,
    CacheHit,
    CacheKey,
    TraceStatus,
    MessagingDestinationName,
    MessagingMessageId,
    ThreadName,
    ThreadId,
    ProfilerId,
    UserCountryCode,
    UserSubregion,

}

A list of supported span tags for tag extraction.

Variants

Release

User

UserID

UserIP

UserUsername

UserEmail

Environment

Transaction

TransactionMethod

TransactionOp

BrowserName

SdkName

SdkVersion

Platform

Mobile

DeviceClass

OsName

Action

AIPipelineGroup

The group of the ancestral span with op ai.pipeline.*

Category

Description

Domain

RawDomain

Group

HttpDecodedResponseContentLength

HttpResponseContentLength

HttpResponseTransferSize

ResourceRenderBlockingStatus

SpanOp

SpanStatus

StatusCode

System

TimeToInitialDisplay

Contributes to Time-To-Initial-Display.

TimeToFullDisplay

Contributes to Time-To-Full-Display.

FileExtension

File extension for resource spans.

MainThread

Span started on main thread.

AppStartType

The start type of the application when the span occurred.

ReplayId

CacheHit

CacheKey

TraceStatus

MessagingDestinationName

MessagingMessageId

ThreadName

ThreadId

ProfilerId

UserCountryCode

UserSubregion

Implementations

impl SpanTagKey

pub fn sentry_tag_key(&self) -> &str

The key used to write this tag into span.sentry_keys.

This key corresponds to the tag key in the snuba span dataset.