relay_event_normalization::span::tag_extraction

Function extract_tags

pub fn extract_tags(
    span: &Span,
    max_tag_value_size: usize,
    initial_display: Option<Timestamp>,
    full_display: Option<Timestamp>,
    is_mobile: bool,
    start_type: Option<&str>,
    span_allowed_hosts: &[String],
    geoip_lookup: &GeoIpLookup,
) -> SentryTags

Writes fields into Span::sentry_tags.

Generating new span data fields is based on a combination of looking at span operations and existing span data fields, and rely on Sentry conventions and heuristics.