relay_event_normalization

Struct BreakdownsConfig

pub struct BreakdownsConfig(pub HashMap<String, BreakdownConfig>);

Configuration for computing breakdowns from data in the event.

Breakdowns are product-defined numbers that are indirectly reported by the client, and are materialized during ingestion. They are usually an aggregation over data present in the event.

Tuple Fields
0: HashMap<String, BreakdownConfig>