Struct relay_event_normalization::LightNormalizationConfig

pub struct LightNormalizationConfig<'a> {

    pub client_ip: Option<&'a IpAddr>,
    pub user_agent: RawUserAgentInfo<&'a str>,
    pub received_at: Option<DateTime<Utc>>,
    pub max_secs_in_past: Option<i64>,
    pub max_secs_in_future: Option<i64>,
    pub transaction_range: Option<Range<UnixTimestamp>>,
    pub max_name_and_unit_len: Option<usize>,
    pub measurements: Option<DynamicMeasurementsConfig<'a>>,
    pub breakdowns_config: Option<&'a BreakdownsConfig>,
    pub normalize_user_agent: Option<bool>,
    pub transaction_name_config: TransactionNameConfig<'a>,
    pub is_renormalize: bool,
    pub device_class_synthesis_config: bool,
    pub enrich_spans: bool,
    pub light_normalize_spans: bool,
    pub max_tag_value_length: usize,
    pub span_description_rules: Option<&'a Vec<SpanDescriptionRule>>,
    pub performance_score: Option<&'a PerformanceScoreConfig>,
    pub geoip_lookup: Option<&'a GeoIpLookup>,
    pub enable_trimming: bool,

}

Configuration for light_normalize_event.

Fields
client_ip: Option<&'a IpAddr>

The IP address of the SDK that sent the event.

When {{auto}} is specified and there is no other IP address in the payload, such as in the request context, this IP address gets added to the user context.

user_agent: RawUserAgentInfo<&'a str>

The user-agent and client hints obtained from the submission request headers.

Client hints are the preferred way to infer device, operating system, and browser information should the event payload contain no such data. If no client hints are present, normalization falls back to the user agent.

received_at: Option<DateTime<Utc>>

The time at which the event was received in this Relay.

This timestamp is persisted into the event payload.

max_secs_in_past: Option<i64>

The maximum amount of seconds an event can be dated in the past.

If the event’s timestamp is older, the received timestamp is assumed.

max_secs_in_future: Option<i64>

The maximum amount of seconds an event can be predated into the future.

If the event’s timestamp lies further into the future, the received timestamp is assumed.

transaction_range: Option<Range<UnixTimestamp>>

The valid time range for transaction events.

This time range should be inferred from storage dependencies, such as metrics storage. Transactions with an end timestamp outside of this time range are dropped as invalid.

max_name_and_unit_len: Option<usize>

The maximum length for names of custom measurements.

Measurements with longer names are removed from the transaction event and replaced with a metadata entry.

measurements: Option<DynamicMeasurementsConfig<'a>>

Configuration for measurement normalization in transaction events.

Has an optional MeasurementsConfig from both the project and the global level. If at least one is provided, then normalization will truncate custom measurements and add units of known built-in measurements.

breakdowns_config: Option<&'a BreakdownsConfig>

Emit breakdowns based on given configuration.

normalize_user_agent: Option<bool>

When Some(true), context information is extracted from the user agent.

transaction_name_config: TransactionNameConfig<'a>

Configuration for sanitizing unparameterized transaction names.

is_renormalize: bool

When Some(true), it is assumed that the event has been normalized before.

This disables certain normalizations, especially all that are not idempotent. The renormalize mode is intended for the use in the processing pipeline, so an event modified during ingestion can be validated against the schema and large data can be trimmed. However, advanced normalizations such as inferring contexts or clock drift correction are disabled.

None equals to false.

device_class_synthesis_config: bool

When true, infers the device class from CPU and model.

enrich_spans: bool

When true, extracts tags from event and spans and materializes them into span.data.

light_normalize_spans: bool

When true, computes and materializes attributes in spans based on the given configuration.

max_tag_value_length: usize

The maximum allowed size of tag values in bytes. Longer values will be cropped.

span_description_rules: Option<&'a Vec<SpanDescriptionRule>>

Configuration for replacing identifiers in the span description with placeholders.

This is similar to transaction_name_config, but applies to span descriptions.

performance_score: Option<&'a PerformanceScoreConfig>

Configuration for generating performance score measurements for web vitals

geoip_lookup: Option<&'a GeoIpLookup>

An initialized GeoIP lookup.

enable_trimming: bool

When Some(true), individual parts of the event payload is trimmed to a maximum size.

See the event schema for size declarations.