Struct relay_event_normalization::StoreProcessor

pub struct StoreProcessor<'a> { /* private fields */ }

The processor that normalizes events for processing and storage.

This processor is a superset of light_normalize_event, that runs additional and heavier normalization steps. These normalizations should ideally be performed on events that are likely to be ingested, after other functionality such as inbound filters have run.

See the fields of StoreConfig for a description of all normalization steps.

Implementations

impl<'a> StoreProcessor<'a>

pub fn new(config: StoreConfig, geoip_lookup: Option<&'a GeoIpLookup>) -> Self

Creates a new normalization processor.

pub fn config(&self) -> &StoreConfig

Returns a reference to the config.