Struct relay_event_normalization::TimestampProcessor

pub struct TimestampProcessor;

Ensures an event’s timestamps are not stale.

Stale timestamps are those that happened before January 1, 1970 UTC. The processor validates the start and end timestamps of an event and returns an error if any of these are stale. Additionally, spans and breadcrumbs with stale timestamps are removed from the event.

The processor checks the timestamps individually and it’s not responsible for decisions that relate timestamps together, including but not limited to:

Ensuring the start timestamp is not later than the end timestamp.

The event finished in the last X days.