Crate relay_event_normalization

Event normalization and processing.

Modules

breakdowns

Computation of breakdowns from event data.

replay

Validation and normalization of [Replay] events.

span

Span normalization logic.

user_agent

Generate context data from user agent and client hints.

utils

Deprecated. Utilities for extracting common event fields.

Structs

BreakdownsConfig

Configuration for computing breakdowns from data in the event.

BuiltinMeasurementKey

Defines a builtin measurement.

ClientHints

The client hint variable names mirror the name of the “SEC-CH” headers.

ClockDriftProcessor

Corrects clock drift based on the sender’s and receivers timestamps.

DynamicMeasurementsConfig

Container for global and project level MeasurementsConfig. The purpose is to handle the merging logic.

GeoIpLookup

A geo ip lookup helper based on maxmind db files.

LightNormalizationConfig

Configuration for light_normalize_event.

MeasurementsConfig

Configuration for measurements normalization.

NormalizeProcessor

The processor that normalizes events for store.

PerformanceScoreConfig

Defines the performance configuration for the project.

PerformanceScoreProfile

Defines a profile for performance score.

PerformanceScoreWeightedComponent

Defines a weighted component for a performance score.

RawUserAgentInfo

A container housing both the user-agent string and the client hint headers.

SpanDescriptionRule

The rule describes how span descriptions should be changed.

SpanDescriptionRuleScope

Object containing transaction attributes the rules must only be applied to.

SpanOperationsConfig

Configuration to define breakdowns based on span operation name.

StoreConfig

Configuration for the StoreProcessor.

StoreProcessor

The processor that normalizes events for processing and storage.

TimeWindowSpan

A time window declared by its start and end timestamp.

TransactionNameConfig

Configuration for sanitizing unparameterized transaction names.

TransactionNameRule

The rule describes how transaction name should be changed.

TransactionsProcessor

Rejects transactions based on required fields.

Enums

BreakdownConfig

Configuration to define breakdown to be generated based on properties and breakdown type.

RedactionRule

Describes what to do with the matched pattern.

Functions

is_high_cardinality_sdk

Returns true if the event’s transaction name is known to contain unsanitized values.

is_valid_platform

Returns true if the given platform string is a known platform identifier.

light_normalize_event

Normalizes data in the event payload.

normalize_breakdowns

Computes breakdowns for an event based on the given configuration.

normalize_ip_addresses

Backfills IP addresses in various places.

normalize_user_agent

Generates context data from client hints or user agent.

normalize_user_agent_info_generic

Low-level version of normalize_user_agent operating on parts.

set_default_transaction_source

Set a default transaction source if it is missing, but only if the transaction name was extracted as a metrics tag. This behavior makes it possible to identify transactions for which the transaction name was not extracted as a tag on the corresponding metrics, because source == null <=> transaction name == null See relay_server::metrics_extraction::transactions::get_transaction_name.

validate_environment

Given a string checks if the environment name is generally valid.

validate_release

Given a string checks if the release is generally valid.

Type Definitions

GeoIpError

An error in the GeoIpLookup.