relay_event_normalization
Module breakdowns

Computation of breakdowns from event data.

Breakdowns are product-defined numbers that are indirectly reported by the client, and are materialized during ingestion. They are usually an aggregation over data present in the event.

Structs

BreakdownsConfig

Configuration for computing breakdowns from data in the event.

SpanOperationsConfig

Configuration to define breakdowns based on span operation name.

TimeWindowSpan

A time window declared by its start and end timestamp.

Enums

BreakdownConfig

Configuration to define breakdown to be generated based on properties and breakdown type.

Functions

normalize_breakdowns

Computes breakdowns for an event based on the given configuration.