relay_event_normalization::request

Function normalize_request

pub fn normalize_request(request: &mut Request)

Normalizes the Request interface.

This function applies the following normalization rules:

The URL is truncated to 2048 characters.

The query string and fragment are extracted into dedicated fields.

The method is normalized to uppercase.

The data is parsed as JSON or urlencoded and put into the data field.

The Content-Type header is parsed and put into the inferred_content_type field.

The Cookie header is parsed and put into the cookies field.