Module relay_event_normalization::span::tag_extraction

Logic for persisting items into span.sentry_tags and span.measurements fields. These are then used for metrics extraction.

Enums

SpanTagKey

A list of supported span tags for tag extraction.

Statics

HTTP_METHOD_EXTRACTOR_REGEX

Regex with a capture group to extract the HTTP method from a string.

Functions

extract_measurements

Copies specific numeric values from span data to span measurements.

extract_segment_span_tags

Extract segment span specific tags and measurements from the event and materialize them into the spans.

extract_span_tags

Extracts tags and measurements from event and spans and materializes them into the spans.

extract_tags

Writes fields into Span::sentry_tags.