Module relay_event_normalization::utils

Deprecated. Utilities for extracting common event fields.

This utility module is being phased out. Functionality in this module should be moved to the specific normalization file requiring this data access.

Constants

MAIN_THREAD_NAME

Allowed value for main thread name.

MAX_DURATION_MOBILE_MS

Maximum length of a mobile span or measurement in milliseconds.

MOBILE_SDKS

Used to decide when to extract mobile-specific tags.

Functions

calculate_cdf_score

Calculates a log-normal CDF score based on a log-normal with a specific p10 and p50

extract_http_status_code

Extracts the HTTP status code.

extract_transaction_op

Returns a normalized op from the given trace context.

get_eventuser_tag

Compute the transaction event’s “user” tag as close as possible to how users are determined in the transactions dataset in Snuba. This should produce the exact same user counts as the user column in Discover for Transactions, barring:

http_status_code_from_span

Extract the HTTP status code from the span data.